# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/policies/search": {
            "get": {
                "description": "Search policies based on various criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policies"
                ],
                "summary": "Search policies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy Type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Company Name",
                        "name": "company",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Policy Id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Policy Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Number of Years",
                        "name": "years",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Policy"
                            }
                        }
                    },
                    "404": {
                        "description": "No policies found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Policy": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "duration_years": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "initial_deposit": {
                    "type": "number"
                },
                "interest": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "term_amount": {
                    "type": "number"
                },
                "terms_per_year": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "user_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}